优化眉毛动作

+++ src/servo_task.h
@@ -1,6 +1,6 @@
 /**
  * @file    servo_task.h
- * @brief   RobotEyes 舵机控制 Task — 平滑非阻塞插值 + 抖动通道
+ * @brief   RobotEyes 舵机控制 Task v10 — 平滑非阻塞插值 + 抖动通道 (int16_t 全链路)
  */
 
 #ifndef SERVO_TASK_H
@@ -18,18 +18,12 @@
 
 extern TaskHandle_t g_servoTaskHandle;
 
+/* v10: 所有角度参数从 int8_t 升级为 int16_t, 防溢出 */
 void servo_task_init(void);
-void servo_set_target(int8_t left_deg, int8_t right_deg);
-
-/* 获取当前舵机目标角度 */
-void servo_get_target(int8_t* left_deg, int8_t* right_deg);
-
-/* 相对当前目标角度进行偏移 (用于摇杆微调) */
-void servo_add_relative(int8_t left_offset, int8_t right_offset);
-
-/* v9.0: 直接注频抖动, 绕过 SERVO_STEP_DEG 缓动限速 (Angry/Panic) */
-void servo_set_jitter(int8_t left_jitter, int8_t right_jitter);
-
+void servo_set_target(int16_t left_deg, int16_t right_deg);
+void servo_get_target(int16_t* left_deg, int16_t* right_deg);
+void servo_add_relative(int16_t left_offset, int16_t right_offset);
+void servo_set_jitter(int16_t left_jitter, int16_t right_jitter);
 void servo_task_run(void* pvParameters);
 
 #endif
